Freeze Fix
Changed '==' to '<=' such that if the versions that are in requirements are not present, it will try and get lower versions that are as close as possible to it. May not work if functionality is not in those versions, but better than not working without that exact version at all.