Metrics and ToppFun Failsafe Update
Implemented failsafe for metrics not working on certain sets of data based on their structure. Now, if a metric is not applicable, it will be NA instead of breaking. Also fixed toppfun breaking on one community and not functioning anymore for the rest, such that as much data as possible can be conserved.